mybatis/src/main/java/examples/boot/mybatis/dao/CityDao.java
package examples.boot.mybatis.dao;

import examples.boot.mybatis.domain.City;
import examples.boot.mybatis.mapper.HotelMapper;
import org.apache.ibatis.session.SqlSession;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Component;

@Component
public class CityDao {
    private final SqlSession sqlSession;

    public CityDao(SqlSession sqlSession) {
        this.sqlSession = sqlSession;
    }

    public City selectCityById(long id) {
        return this.sqlSession.selectOne("selectCityById", id);
    }
}